# public/openapi/axelot-api.yaml
openapi: 3.1.0
info:
  title: Axelot API
  version: 1.0.0
  description: |
    Collaborative story writing platform API with real-time editing,
    AI-powered text augmentation, and social discovery features.

    Authentication uses NextAuth v5 with JWT sessions. Some endpoints
    require a Firebase ID token obtained by exchanging the custom token
    from the NextAuth session via the Firebase client SDK.
  contact:
    name: Royce Mathew & Sunny Patel
    email: sunnypatel124555@gmail.com
    url: https://github.com/royce-mathew/axelot
  license:
    name: AGPL-3.0
    url: https://github.com/royce-mathew/axelot/blob/master/LICENSE
servers:
  - url: https://axelot.io
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/stories/discover:
    get:
      tags: [Stories]
      operationId: discoverStories
      summary: Discover public stories
      description: |
        Returns a paginated list of public, non-archived stories.
        Three discovery modes are available:

        - **all**: Sorted by `lastUpdated` descending. Shows the most
          recently active stories.
        - **foryou**: Sorted by `trendingScore` descending. Powered by
          the trending algorithm that factors in views, recency, and age.
        - **fresh**: Sorted by `created` descending. Shows the newest
          stories first.

        Results are cached server-side for performance. Cache durations
        vary by mode: 10 minutes for `all` and `foryou`, 2 minutes for
        `fresh`.

        An optional search parameter filters results by title (case-
        insensitive, client-side substring match on the first 100
        results).
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum:
              - all
              - foryou
              - fresh
            default: all
          description: |
            Discovery mode controlling sort order.
          example: foryou
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: |
            Zero-indexed page number for pagination.
          example: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: |
            Number of stories per page.
          example: 20
        - name: search
          in: query
          schema:
            type: string
          description: |
            Optional search term to filter stories by title.
            Searches the first 100 results client-side.
          example: science fiction
      responses:
        "200":
          description: Paginated list of public stories.
          headers:
            Cache-Control:
              schema:
                type: string
              description: |
                Server-side cache duration.
                `s-maxage=600, stale-while-revalidate` for all/foryou,
                `s-maxage=120, stale-while-revalidate` for fresh.
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Story"
                  total:
                    type: integer
                    description: |
                      Total matching stories. Only included when a
                      search term is provided.
                  hasMore:
                    type: boolean
                    description: Whether more pages are available.
              examples:
                stories:
                  summary: Page of discovered stories
                  value:
                    stories:
                      - id: abc123
                        title: The Last Frontier
                        owner: user456
                        authorNames: [Alice, Bob]
                        slug: the-last-frontier
                        created: "2025-12-01T10:00:00Z"
                        lastUpdated: "2025-12-15T14:30:00Z"
                        tags: [scifi, adventure]
                        viewCount: 142
                        trendingScore: 0.873
                        isPublic: true
                        isArchived: false
                        preview: "The ship emerged from hyperspace into a star system no human had ever visited..."
                    hasMore: true
        "500":
          description: Firestore query failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                failure:
                  value:
                    error: Failed to fetch stories
  /api/stories/trending:
    get:
      tags: [Stories]
      operationId: trendingStories
      summary: Get trending stories
      description: |
        Returns public, non-archived stories sorted by their trending score
        in descending order. Scores are computed by a background job that
        runs daily and considers:

        - **View count** (logarithmic scaling to prevent viral domination)
        - **Recency boost** (linear decay over 24 hours since last update)
        - **Age penalty** (exponential decay based on creation age)
        - **Activity boost** (extra weight for very recently updated stories)

        The formula:
        `score = log10(views + 1) * 4 + recencyBoost * 10 - agePenalty * 0.1`

        Only stories with `isPublic: true` and `isArchived: false` are
        included. Results are cached for 10 minutes.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Zero-indexed page number.
          example: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Stories per page.
          example: 20
      responses:
        "200":
          description: Trending stories sorted by score.
          headers:
            Cache-Control:
              schema:
                type: string
              description: "`s-maxage=600, stale-while-revalidate`"
          content:
            application/json:
              schema:
                type: object
                properties:
                  stories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Story"
                  hasMore:
                    type: boolean
              examples:
                trending:
                  summary: Top trending stories
                  value:
                    stories:
                      - id: abc123
                        title: The Last Frontier
                        owner: user456
                        authorNames: [Alice]
                        slug: the-last-frontier
                        created: "2025-12-01T10:00:00Z"
                        lastUpdated: "2025-12-15T14:30:00Z"
                        tags: [scifi]
                        viewCount: 1420
                        trendingScore: 2.15
                        isPublic: true
                        isArchived: false
                        preview: "The ship emerged from hyperspace..."
                    hasMore: true
        "500":
          description: Firestore query failure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    firebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A Firebase ID token obtained by exchanging the custom token
        from the NextAuth session using the Firebase client SDK.

        Flow:
        1. Sign in via NextAuth (OAuth or credentials)
        2. NextAuth session callback generates a Firebase custom token
        3. Client exchanges custom token via `signInWithCustomToken()`
        4. Client calls `getIdToken()` to get the Firebase ID token
        5. Include in requests: `Authorization: Bearer <firebase-id-token>`
    cronSecret:
      type: http
      scheme: bearer
      description: |
        A shared secret configured in the `CRON_SECRET` environment
        variable. Used to authenticate cron job requests from Vercel
        or manual triggers.

        Include in requests: `Authorization: Bearer <cron-secret>`
  schemas:
    Error:
      type: object
      description: Standard error response.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: Missing required parameter
    Story:
      type: object
      description: |
        A collaborative story document. Stories have ownership, access
        controls, metadata, engagement metrics, and content stored as
        Yjs-encoded binary data in Firestore.
      properties:
        id:
          type: string
          description: Firestore document ID.
          example: abc123def456
        owner:
          type: string
          description: Firebase user ID of the story owner.
          example: user789
        authorNames:
          type: array
          items:
            type: string
          description: |
            Display names of all collaborators (owner + write access users).
            Denormalized for performance to avoid joins at read time.
          example: [Alice Johnson, Bob Smith]
        title:
          type: string
          description: Story title set by the owner.
          example: The Last Frontier
        description:
          type: string
          nullable: true
          description: Optional story description or subtitle.
          example: A tale of exploration beyond known space.
        slug:
          type: string
          description: |
            URL-friendly slug derived from the title. Auto-generated
            when the title changes.
          example: the-last-frontier
        created:
          type: string
          format: date-time
          description: When the story was first created.
          example: "2025-12-01T10:00:00Z"
        lastUpdated:
          type: string
          format: date-time
          description: When the story content or metadata was last modified.
          example: "2025-12-15T14:30:00Z"
        lastUpdatedBy:
          type: string
          description: Firebase user ID of the last person to edit.
          example: user789
        tags:
          type: array
          items:
            type: string
          description: User-defined tags for categorization.
          example: [scifi, adventure, collaborative]
        isPublic:
          type: boolean
          description: |
            Whether the story is visible in the discovery feed and
            accessible to anyone with the link.
          example: true
        isArchived:
          type: boolean
          description: |
            Archived stories are hidden from the discovery feed but
            still accessible to the owner and collaborators.
          default: false
          example: false
        viewCount:
          type: integer
          description: |
            Total number of authenticated views. Incremented by the
            `/api/document/view` endpoint.
          default: 0
          example: 142
        lastViewed:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the most recent view.
          example: "2025-12-16T09:15:00Z"
        trendingScore:
          type: number
          format: float
          description: |
            Algorithmic trending score computed from views, recency,
            and age. Updated by the cron job. Higher is more trending.
          default: 0
          example: 2.15
        trendingLastComputed:
          type: string
          format: date-time
          nullable: true
          description: When the trending score was last recomputed.
          example: "2025-12-16T00:00:00Z"
        preview:
          type: string
          description: |
            Plain text excerpt extracted from the Yjs-encoded story
            content. Generated at query time, not stored in Firestore.
          example: "The ship emerged from hyperspace into a star system no human had ever visited..."
        readAccess:
          type: array
          items:
            type: string
          description: |
            Firebase user IDs with read-only access. These users can
            view the story even when it is private.
          example: [viewer1, viewer2]
        writeAccess:
          type: array
          items:
            type: string
          description: |
            Firebase user IDs with write access. These users can
            collaboratively edit the story in real time and appear
            in the `authorNames` array.
          example: [editor1]
...